ai-service/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0

langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
langgraph>=0.2.0,<1.0.0

pypdf>=4.3.0,<6.0.0
Pillow>=10.4.0,<12.0.0
rapidocr-onnxruntime>=1.2.0,<1.3.0
pymongo>=4.8.0,<5.0.0
